Added option to reset the infinite loader
When using the infinite scroll there is no ability to reset the scroll element. This can be useful when you are using the infinite scroll list to show results based on filters. If you apply filter A and scroll till the end the message no more data is shown. But if you also apply filter B then you will reset the list and also the messages that there are either no results or no more data.