Fix #43: Infinite loading tables
`this.$el.offsetTop` will always evaluate to 0 when `this.$el` is inside a `td` (but the scrolling element is a `table`).

This PR fixes the issue by explicitly calculating the top offset of `this.$el`, relative to the actual scrolling element, and not just the direct parent.

I tried to see if `this.$el.offsetTop` will be updated by just setting CSS `position` to `relative` on `tr`, `td`, and `this.$el`, and all combinations, but no luck there. The performance hit of this PR  will be an extra `getBoundingClientRect` call for `this.$el`.